Authorization

+++ backend/routes/posts.js
@@ -38,7 +38,8 @@ router.post(
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: imageURL + '/images/' + req.file.filename
+            imagePath: imageURL + '/images/' + req.file.filename,
+            creator: req.userData.userId
         });
         // Mongoose creates query for db to create a new entry to MongoDB
         post.save().then(createdPost => {
@@ -70,11 +71,15 @@ router.put(
             _id: req.body.id,
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
-        Post.updateOne({ _id: req.params.id }, post).then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Update Successful' });
+        Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+            if (result.modifiedCount > 0) {
+                res.status(200).json({ message: 'Update Successful' });
+            } else {
+                res.status(401).json({ message: 'User not Authorized' });
+            }
         });
     });
 
@@ -125,8 +130,12 @@ router.delete(
     '/:id',
     checkAuth,
     (req, res, next) => {
-        Post.deleteOne({ _id: req.params.id }).then(result => {
-            res.status(200).json({ message: 'Post Deleted' });
+        Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+            if (result.deletedCount > 0) {
+                res.status(200).json({ message: 'Delete Successful' });
+            } else {
+                res.status(401).json({ message: 'User not Authorized' });
+            }
         });
     });
